+++ config/routes.rb
@@ -26,7 +26,8 @@ Rails.application.routes.draw do
           'authenticate#authenticate'
       end
 
-      get "/roles/:account/:kind/*identifier" => "roles#memberships", :constraints => QueryParameterActionRecognizer.new("all")
+      get "/roles/:account/:kind/*identifier" => "roles#all_memberships", :constraints => QueryParameterActionRecognizer.new("all")
+      get "/roles/:account/:kind/*identifier" => "roles#direct_memberships", :constraints => QueryParameterActionRecognizer.new("memberships")
 
       get "/roles/:account/:kind/*identifier" => "roles#show"