+++ app/models/role.rb
@@ -108,4 +108,8 @@ class Role < Sequel::Model
   def all_roles
     Role.from(Sequel.function(:all_roles, id))
   end
+
+  def ancestor_of? role
+    Role.from(Sequel.function(:is_role_ancestor_of, id, role.id)).first[:is_role_ancestor_of]
+  end
 end