+++ CHANGELOG.md
@@ -25,6 +25,8 @@ permission on them.
 - Experimental audit querying engine mounted at /audit. It can be configured to work with
 an external audit database by using config.audit_database configuration entry.
 
+- API endpoints for granting and revoking role membership
+
 ## [0.4.0] - 2018-04-10
 ### Added
 - Policy changes now generate audit log messages. These can optionally be generated in RFC5424