+++ CHANGELOG.md
@@ -29,6 +29,8 @@ an external audit database by using config.audit_database configuration entry.
 
 - API endpoint for the role graph
 
+- Paging parameters (`offset` and `limit`) for audit API endpoints
+
 ## [0.4.0] - 2018-04-10
 ### Added
 - Policy changes now generate audit log messages. These can optionally be generated in RFC5424